add imagemin plugin and optimize preset also remove pngquant and svgo
