+++ packages/plugin-imagemin/README.md
@@ -1,0 +1,19 @@
+# magefront-plugin-imagemin
+
+Optimize images with [imagemin](https://github.com/imagemin/imagemin).
+
+## Install
+
+    npm i magefront-plugin-imagemin
+
+## Usage
+
+```js
+import imagemin from 'magefront-plugin-imagemin'
+
+export default {
+  plugins: [imagemin()],
+}
+```
+
+See the [documentation](https://ubermanu.github.io/magefront/#/plugins/imagemin) for more information.
