--- packages/plugin-pngquant/CHANGELOG.md
@@ -1,65 +1,0 @@
-# magefront-plugin-pngquant
-
-## 1.2.2
-
-### Patch Changes
-
-- magefront@1.2.2
-
-## 1.2.1
-
-### Patch Changes
-
-- magefront@1.2.1
-
-## 1.2.0
-
-### Minor Changes
-
-- f8da0cb: Set a proper (peer) dependency version in packages
-
-### Patch Changes
-
-- Updated dependencies [f8da0cb]
-- Updated dependencies [f8da0cb]
-  - magefront@1.2.0
-
-## 1.1.0
-
-### Minor Changes
-
-- 4cc6bd1: Changeset is struggling with peer dependencies.
-  Set the magefront peer dependency to `any`.
-
-### Patch Changes
-
-- Updated dependencies [9d2b2b5]
-  - magefront@1.1.0
-
-## 1.0.0
-
-### Major Changes
-
-- Release 1.0
-
-### Minor Changes
-
-- 129c7ba: Loosen the magefront peer dependency version
-
-### Patch Changes
-
-- 886c513: Update dependencies to latest version
-- Updated dependencies [3b70273]
-- Updated dependencies [886c513]
-- Updated dependencies
-  - magefront@1.0.0
-
-## 0.4.0
-
-### Minor Changes
-
-- 8784523: Set magefront peer dependency version to 'any' to avoid warnings
-
-### Patch Changes
-
-- magefront@0.24.3
