--- packages/plugin-pngquant/README.md
@@ -1,19 +1,0 @@
-# magefront-plugin-pngquant
-
-Optimize PNG files with [pngquant](https://pngquant.org/).
-
-## Install
-
-    npm i magefront-plugin-pngquant
-
-## Usage
-
-```js
-import pngquant from 'magefront-plugin-pngquant'
-
-export default {
-  plugins: [pngquant()],
-}
-```
-
-See the [documentation](https://ubermanu.github.io/magefront/#/plugins/pngquant) for more information.
