--- packages/plugin-svgo/README.md
@@ -1,19 +1,0 @@
-# magefront-plugin-svgo
-
-Optimize SVG files with [SVGO](https://github.com/svg/svgo).
-
-## Install
-
-    npm i magefront-plugin-svgo
-
-## Usage
-
-```js
-import svgo from 'magefront-plugin-svgo'
-
-export default {
-  plugins: [svgo()],
-}
-```
-
-See the [documentation](https://ubermanu.github.io/magefront/#/plugins/svgo) for more information.
