--- packages/plugin-svgo/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "magefront-plugin-svgo",
-  "version": "1.2.2",
-  "keywords": [
-    "magefront",
-    "magento",
-    "theme",
-    "plugin"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ubermanu/magefront.git",
-    "directory": "packages/plugin-svgo"
-  },
-  "license": "MIT",
-  "author": "Emmanuel Vodor <manu@paca.happy-dev.fr>",
-  "type": "module",
-  "exports": {
-    ".": {
-      "import": "./dist/plugin.mjs",
-      "require": "./dist/plugin.cjs",
-      "types": "./dist/plugin.d.ts"
-    }
-  },
-  "main": "dist/plugin.cjs",
-  "module": "dist/plugin.mjs",
-  "types": "dist/plugin.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "unbuild"
-  },
-  "dependencies": {
-    "fast-glob": "^3.3.1",
-    "svgo": "^3.0.2"
-  },
-  "peerDependencies": {
-    "magefront": "workspace:^"
-  }
-}
