+++ packages/preset-optimize/README.md
@@ -1,0 +1,19 @@
+# magefront-preset-optimize
+
+Regroup a collection of `imagemin` plugins to optimize images.
+
+## Install
+
+    npm i magefront-preset-optimize
+
+## Usage
+
+```js
+import optimizePreset from 'magefront-preset-optimize'
+
+export default {
+  presets: [optimizePreset()],
+}
+```
+
+See the [documentation](https://ubermanu.github.io/magefront/#/presets/optimize) for more information.
