+++ packages/preset-optimize/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "magefront-preset-optimize",
+  "version": "1.0.0",
+  "keywords": [
+    "magefront",
+    "magento",
+    "theme",
+    "preset",
+    "optimize",
+    "image"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ubermanu/magefront.git",
+    "directory": "packages/preset-optimize"
+  },
+  "license": "MIT",
+  "author": "Emmanuel Vodor <manu@paca.happy-dev.fr>",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": "./dist/preset.mjs",
+      "require": "./dist/preset.cjs",
+      "types": "./dist/preset.d.ts"
+    }
+  },
+  "main": "dist/preset.cjs",
+  "module": "dist/preset.mjs",
+  "types": "dist/preset.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "unbuild"
+  },
+  "dependencies": {
+    "imagemin-gifsicle": "^7.0.0",
+    "imagemin-jpegtran": "^7.0.0",
+    "imagemin-pngquant": "^9.0.2",
+    "imagemin-svgo": "^10.0.1",
+    "imagemin-webp": "^8.0.0",
+    "magefront-plugin-imagemin": "workspace:^"
+  },
+  "devDependencies": {
+    "@types/imagemin-gifsicle": "^7.0.1",
+    "@types/imagemin-jpegtran": "^5.0.1",
+    "@types/imagemin-svgo": "^10.0.2"
+  },
+  "peerDependencies": {
+    "magefront": "workspace:^"
+  }
+}
